ledger/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.7.0
jsonschema==4.22.0
PyNaCl==1.5.0
